feat(api): Add API to repeat last selected item
