+++ lua/legendary/ui/init.lua
@@ -7,29 +7,6 @@ local Format = require('legendary.ui.format')
 local Executor = require('legendary.api.executor')
 local Log = require('legendary.log')
 
-local function update_item_frecency_score(item)
-  if Config.sort.frecency ~= false then
-    if require('legendary.api.db').is_supported() then
-      Log.trace('Updating scoring data for selected item.')
-      local DbClient = require('legendary.api.db.client').init()
-      -- if bootstrapping fails, bail
-      if not require('legendary.api.db').is_supported() then
-        Log.debug(
-          'Config.sort.frecency is enabled, but sqlite is not available or database could not be opened, '
-            .. 'frecency is automatically disabled.'
-        )
-        return
-      end
-      DbClient.update_item_score(item)
-    else
-      Log.debug(
-        'Config.sort.frecency is enabled, but sqlite is not available or database could not be opened, '
-          .. 'frecency is automatically disabled.'
-      )
-    end
-  end
-end
-
 ---@class LegendaryUi
 ---@field select fun(opts:LegendaryFindOpts)
 local M = {}
@@ -90,17 +67,12 @@ local function select_inner(opts, context, itemlist)
       return
     end
 
-    local ok, err = pcall(update_item_frecency_score, selected)
-    if not ok then
-      Log.error('Failed to update frecency score: %s', err)
-    end
-
-    State.most_recent_item = selected
     if Toolbox.is_itemgroup(selected) then
       return select_inner(opts, context, selected.items)
     end
 
     Log.trace('Preparing to execute selected item')
+    State.most_recent_item = selected
     Executor.exec_item(selected, context)
   end)
 end
